# Add parallel computing interface for farm simulation and yaw optimization