--- a/floris/tools/__init__.py
+++ b/floris/tools/__init__.py
@@ -38,6 +38,7 @@ Examples:
 
 from .floris_interface import FlorisInterface
 from .floris_interface_legacy_reader import FlorisInterfaceLegacyV2
+from .parallel_computing_interface import ParallelComputingInterface
 from .visualization import visualize_cut_plane, visualize_quiver, plot_turbines_with_fi, plot_rotor_values
 from .wind_rose import WindRose
 from .uncertainty_interface import UncertaintyInterface