… docs/examples.md
… examples/12_optimize_yaw_in_parallel.py