<!-- Is this pull request ready to be merged? -->
<!-- i.e. tests pass or are expected to fail; all development is finished; appropriate documentation is included. -->
<!-- If not but opening the pull request will facilitate development, make it a "draft" pull request -->

**Feature or improvement description**
This PR adds a parallel computing interface at the tools level to enable parallel computing of turbine powers, wind farm powers, and yaw optimizations. It does not change anything to the internal FLORIS code, which makes it a much easier integration than #403. I have been using this internally for quick parallel computing solutions using `multiprocessing` for single-node jobs/local parallelization, but this class also supports `mpi4py` for multinode calculations like FLORIS v2 used to.

**Related issue, if one exists**
#401 


**Impacted areas of the software**
FLORIS tools library

**Additional supporting information**
The Parallel Computing Interface class is built similarly to the UncertaintyInterface class. It can be loaded up and then interacted with like a regular `FlorisInterface` object, but then the underlying code performs the computations in parallel rather than in serial. In larger problems and especially in yaw optimizations, this can provide significant speed-up.

**Test results, if applicable**
This PR adds an example file that calculates the optimal yaw angles and the resulting AEP uplift using parallel processing, "12_optimize_yaw_in_parallel.py".

With a maximum of one core, the total time for the yaw angle optimization (using SR) and for the AEP calculation are, respectively:
```
==============================================
[Yaw optimization with SR]
Total time spent for parallel calculation (1 workers): 280.834 s
  Time spent in parallel preprocessing: 0.001 s
  Time spent in parallel loop execution: 280.831 s.
  Time spent in parallel postprocessing: 0.002 s
==============================================
[AEP calculation]
Total time spent for parallel calculation (1 workers): 35.606 s
  Time spent in parallel preprocessing: 0.007 s
  Time spent in parallel loop execution: 35.596 s.
  Time spent in parallel postprocessing: 0.004 s
```

And in parallel with 16 cores on my laptop:
```
==============================================
[Yaw optimization with SR]
Total time spent for parallel calculation (16 workers): 87.727 s
  Time spent in parallel preprocessing: 0.003 s
  Time spent in parallel loop execution: 87.720 s.
  Time spent in parallel postprocessing: 0.004 s
==============================================
[AEP calculation]
Total time spent for parallel calculation (16 workers): 11.781 s
  Time spent in parallel preprocessing: 0.012 s
  Time spent in parallel loop execution: 11.763 s.
  Time spent in parallel postprocessing: 0.006 s
```

And the example produces the results:
![image](https://user-images.githubusercontent.com/22119448/211356095-cc5ed06e-17b0-4de7-8d22-8d957b8c24cc.png)

![image](https://user-images.githubusercontent.com/22119448/211356150-0d0bc95c-c11d-40d4-9995-bb4a217142fd.png)

![image](https://user-images.githubusercontent.com/22119448/211356196-b217f102-08c1-430e-9990-e2080e36f4c1.png)

<!-- Release checklist:
- Update the version in
    - [ ] README.rst
    - [ ] docs/index.rst
    - [ ] floris/VERSION
- [ ] Verify readthedocs builds correctly
- [ ] Create a tag in the NREL/FLORIS repository
-->