@mr483 2023-01-13
Hi @Bartdoekemeijer, I was testing out this interface and it works nicely! One question: have you done any sensitivity study on performance improvement with number of cores used and/or number of condition splits? I'm finding on my machine (16 cores) that the performance of running fi_parallel.get_turbine_powers() for 400 turbines does not necessarily show the scaling up of improvement as I would expect:

| Cores (max_workers) | Runtime (s) | Improvement |
| :---         |     :---:      |          ---: |
| 1   | 890  | -  |
| 4   |  270 | 3.3     |
| 8   | 245       | 3.63      |
| 12   | 230       | 3.87      |
| 16   | 238       |  3.74     |

In each case, I confirmed that selecting to split by either just WD (e.g., 8 workers, 8 wd splits, 1 ws split), just WS, or some combination of WS/WD which multiplies to the number of workers (e.g., 4 wd splits, 2 ws splits) didn't make a meaningful difference, as expected.