@Bartdoekemeijer 2023-01-13
> Hi @Bartdoekemeijer, I was testing out this interface and it works nicely! One question: have you done any sensitivity study on performance improvement with number of cores used and/or number of condition splits? I'm finding on my machine (16 cores) that the performance of running fi_parallel.get_turbine_powers() for 400 turbines, 30 wind speeds, 60 wind directions, does not necessarily show the scaling up of improvement as I would expect: 
> ...

Hi @mr483, thank you for testing out this PR! I have found similar things as you did: performance does not always increase with more cores. I think this is because for large wind farms and many wind directions/speeds, the limiting factor in calculations is the available memory rather than CPU power. Hence, more cores then does not speed up the code, but does add additional overhead, so the net result is a slow-down of the code. This is especially apparent with operations like `get_farm_power`, `get_turbine_power` and `get_farm_AEP`.
If you would use this parallel computing interface for yaw optimization, then the benefit of parallel computing becomes a lot clearer. I think the yaw optimization is much more CPU bottlenecked.