@paulf81 2023-02-01
hi @Bartdoekemeijer  thanks a lot of this PR!  I was wondering, we're planning to do some focused development of FLORIS next week, do you think this PR is ready to merge?