@Bartdoekemeijer 2023-02-01
Yes ready to merge if you ask me! I have been very actively using it on different systems and with and without MPI. It could use some documentation but that can be merged in later.