@rafmudaf 2023-02-01
Please include example in the listing here: https://github.com/NREL/floris/blob/develop/docs/examples.md