@Bartdoekemeijer 2023-02-03
> Please include example in the listing here: https://github.com/NREL/floris/blob/develop/docs/examples.md

Done!